Reorganize Packages / Bump to .NET 6.0
