+++ WGCDynamics/model/FeedBack.cs
@@ -16,7 +16,7 @@
  */
 using System;
 
-namespace WGCDynamics.model
+namespace WebGate.Dynamics.Model
 {
     #pragma warning disable IDE1006 // Benennungsstile
     public class FeedBack
