+++ WGCDynamics/model/TransactionCurrency.cs
@@ -15,7 +15,7 @@
  *
  */
 using System;
-namespace WGCDynamics.model
+namespace WebGate.Dynamics.Model
 {
     #pragma warning disable IDE1006 // Benennungsstile
     public class TransactionCurrency
